fix(replay): extract mask settle analysis and halve the fidelity drift budget
